# simpler and more efficient implementation of unused-parameter