--- a/test/unused-param_test.go
+++ b/test/unused-param_test.go
@@ -9,3 +9,10 @@ import (
 func TestUnusedParam(t *testing.T) {
 	testRule(t, "unused-param", &rule.UnusedParamRule{})
 }
+
+func BenchmarkUnusedParam(b *testing.B) {
+	var t *testing.T
+	for i := 0; i <= b.N; i++ {
+		testRule(t, "unused-param", &rule.UnusedParamRule{})
+	}
+}