@mgechev 2019-04-21
>now a parameter is considered used if it is referenced somewhere in the function body (this includes LHS of assignments)

This makes even more sense to me. If we reassign a reference, this shouldn't be considered as not unused.

That's a great catch @chavacava!